% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-attribs.R
\name{board_classic}
\alias{board_classic}
\title{Is the board using classic game mode?}
\usage{
board_classic(board)
}
\arguments{
\item{board}{A \code{lightsout} board object}
}
\description{
In classic mode, pressing a light will toggle it and its adjacent neighbours only.
If non-classic mode, pressing a light will toggle the entire row and column of the pressed light.
}
\keyword{internal}

