% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-attribs.R
\name{board_entries<-}
\alias{board_entries<-}
\title{Set the board entries}
\usage{
board_entries(board) <- value
}
\arguments{
\item{board}{A \code{lightsout} board object}

\item{value}{The new value to set the entries to}
}
\description{
Set the board entries
}
\keyword{internal}

