% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-attribs.R
\name{board_size}
\alias{board_size}
\title{Get the board size (number of rows/columns)}
\usage{
board_size(board)
}
\arguments{
\item{board}{A \code{lightsout} board object}
}
\description{
Get the board size (number of rows/columns)
}
\keyword{internal}

