% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board-attribs.R
\name{board_toggle_matrix}
\alias{board_toggle_matrix}
\title{Get the toggle matrix used to solve the board using linear algebra}
\usage{
board_toggle_matrix(board)
}
\arguments{
\item{board}{A \code{lightsout} board object}
}
\description{
A lightsout game can be solved using Gaussian elimination. A special matrix
\code{A} needs to be used to solve \code{Ax=b}, where \code{b} is the current
board configuration and \code{x} is the solution. This function returns
the matrix \code{A}.
}
\keyword{internal}

