% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsout-package.R
\docType{package}
\name{lightsout-package}
\alias{lightsout}
\alias{lightsout-package}
\title{lightsout: Implementation of the 'Lights Out' Puzzle Game}
\description{
Lights Out is a puzzle game consisting of a grid of lights that are either on or off. Pressing any light will toggle it and its adjacent lights. The goal of the game is to switch all the lights off. This package provides an interface to play the game on different board sizes, both through the command line or with a visual application. Puzzles can also be solved using the automatic solver included. View a demo online at \url{https://daattali.com/shiny/lightsout/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/lightsout}
  \item \url{https://daattali.com/shiny/lightsout/}
  \item Report bugs at \url{https://github.com/daattali/lightsout/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

}
\keyword{internal}
