% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{print.lightsout_solution}
\alias{print.lightsout_solution}
\title{Print a lightsout board solution}
\usage{
\method{print}{lightsout_solution}(x, ...)
}
\description{
Print a lightsout board solution
}
\keyword{internal}

