% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{solve_board}
\alias{solve_board}
\title{Solve a lightsout board}
\usage{
solve_board(board)
}
\description{
Given a 5x5 lightsout game board, find the minimum number of clicks
to solve the board.
}
\details{
The input must be a 5x5 matrix of 0s and 1s.

The return value is a matrix with the same dimensions as the game board, with
a 1 in every position that requires a click. The order of clicks doesn't matter.

Note that a solution will ALWAYS be given, even if the board is not actually
solveable. Therefore, you should check the solution by "clicking" all the 1s
in the answer and see if the board actually gets solved. If not, then that means
there is no solution to the given board.
}

